Addded doxy comments to material

+++ Model_Loading_Library/src/material.cpp
@@ -52,7 +52,7 @@ string Material::getName()
 {
   return this->name;
 }
-double Material::getID()
+int Material::getID()
 {
   return this->id;
 }
